backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pypdf
chromadb
langchain
langchain-community
SpeechRecognition
google-generativeai
python-dotenv
